新增帮助说明、调整执行路径

+++ my_code/sh/excel2csv.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 #Description: csv excel 格式转换
-tmp="$HOME/Documents/tmp.xlsx"
-
-cp $1 $tmp
-python ${CODE_HOME}py/excel2csv.py $tmp $2 $3
-rm $tmp
+if [ $1 = h ];then
+    echo "1:input_file 2:output_file 3:sheet_num[0,1,2,...]"
+else
+    tmp="$HOME/Documents/tmp.xlsx"
+    cp $1 $tmp
+    python ${CODE_HOME}my_code/py/excel2csv.py $tmp $2 $3
+    rm $tmp
+fi
